fix: Update code to interface with Chia 2.0.1+
* chore: Updating changelog for 1.0.28

* chore: Updating changelog for 1.0.29

* chore: Updating changelog for 1.0.30

* chore: Updating changelog for 1.0.31

* chore: Updating changelog for 1.0.32

* chore: Updating changelog for 1.0.33

* Update chia-blockchain to 2.1.0 and major updates to precommit and mypy cleanup

* Poetry updates

* ci: add badge

* ci: add python tests

* ci: make venv adjustments

* ci: adjust testing venvs

* ci: adjust venvs for testing

* ci: tests in ci venvs

* Update pytest.ini to ignore sqlalchemy error

* Update tests to not run TestActivities as a workaround

* set MODE to dev for test run

* setup env for tests

* setup env for tests

* Update tests.yaml

* Update tests.yaml

* Update tests.yaml

* Update tests.yaml

* Update tests.yaml

* Update tests.yaml

* Update tests.yaml

* Update tests.yaml

* Update tests.yaml

* Some general cleanup

* mock code that required a running node

* run all tests

* override anyio to earlier version for tests

* workaround incompatability with anyio 4 and some test packages

* Add nft_wallet stuff to pyinstaller.spec

* Add chia hidden imports to pyinstaller

* Add bunch of logging to get_activity

* Some more logging

* print to stdout for now

* pretty print the dicts

* Fix problem with wrong variable name

---------

+++ app/api/v1/keys.py
@@ -6,6 +6,7 @@ from chia.rpc.wallet_rpc_client import WalletRpcClient
 from chia.types.blockchain_format.sized_bytes import bytes32
 from chia.util.bech32m import decode_puzzle_hash, encode_puzzle_hash
 from chia.util.byte_types import hexstr_to_bytes
+from chia.util.ints import uint32
 from chia.wallet.derive_keys import (
     master_sk_to_wallet_sk,
     master_sk_to_wallet_sk_unhardened,
@@ -31,7 +32,6 @@ async def get_key(
     prefix: str = "bls1238",
     wallet_rpc_client: WalletRpcClient = Depends(deps.get_wallet_rpc_client),
 ):
-
     fingerprint: int = await wallet_rpc_client.get_logged_in_fingerprint()
 
     result: Dict = await wallet_rpc_client.get_private_key(fingerprint)
@@ -40,10 +40,10 @@ async def get_key(
 
     wallet_secret_key: PrivateKey
     if hardened:
-        wallet_secret_key = master_sk_to_wallet_sk(secret_key, derivation_index)
+        wallet_secret_key = master_sk_to_wallet_sk(secret_key, uint32(derivation_index))
     else:
         wallet_secret_key = master_sk_to_wallet_sk_unhardened(
-            secret_key, derivation_index
+            secret_key, uint32(derivation_index)
         )
 
     wallet_public_key: G1Element = wallet_secret_key.get_g1()
@@ -63,7 +63,6 @@ async def get_key(
 async def parse_key(
     address: str,
 ):
-
     try:
         puzzle_hash: bytes = decode_puzzle_hash(address)
     except ValueError:
